requirements.txt
langchain
flask
pypdf
python-dotenv
pinecone
langchain-pinecone
langchain-huggingface
langchain_community
langchain_openai
langchain_experimental
-e .